backend-local/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
requests==2.32.3
yfinance==0.2.55